requirements.txt
altgraph==0.17
backcall==0.2.0
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.9
flake8==3.9.2
importlib-metadata==4.5.0
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
parso==0.7.1
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
zipp==3.4.1
